# swagger/Schema.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "WebApiTest",
      "version": "v1"
    },

    "servers":[
      {
        "url": "https://localhost:5001/"
      }
            
    ],
    "paths": {
      "/api/Books": {
        "get": {
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDto"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDto"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookDto"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Books"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewBookDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewBookDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/NewBookDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Books/{id}": {
        "get": {
          "tags": [
            "Books"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "Books"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBookDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBookDto"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBookDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/BookDto"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Books"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/Books/consumeApi": {
        "get": {
          "tags": [
            "Books"
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookFApi"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookFApi"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BookFApi"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "BookDto": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "pageCount": {
              "type": "integer",
              "format": "int32"
            },
            "excerpt": {
              "type": "string",
              "nullable": true
            },
            "publishDate": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "NewBookDto": {
          "required": [
            "author",
            "name"
          ],
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "author": {
              "type": "string"
            }
          },
          "additionalProperties": false
        },
        "UpdateBookDto": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "nullable": true
            },
            "author": {
              "type": "string",
              "nullable": true
            },
            "genre": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "BookFApi": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "title": {
              "type": "string",
              "nullable": true
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "pageCount": {
              "type": "integer",
              "format": "int32"
            },
            "excerpt": {
              "type": "string",
              "nullable": true
            },
            "publishDate": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        }
      }
    }
  }